+++ Lib/Player/Inc/flac_reader.h
@@ -11,6 +11,6 @@ typedef struct {
 
 FlacReader *create_flac_reader(Flac *flac);
 void free_flac_reader(FlacReader *reader);
-unsigned read_flac(FlacReader *reader, uint16_t  *buffer, unsigned size);
+unsigned read_flac(FlacReader *reader, uint8_t  *buffer, unsigned size);
 
 #endif //STM32_FLAC_PLAYER_FLAC_READER_H
